objectvetype  admin BE

+++ Models/ObjectiveModel.cs
@@ -13,7 +13,7 @@ namespace TravelSBE.Models
         public double Latitude { get; set; }
         public double Longitude { get; set; }
         public string? City { get; set; }
-        public int? MedieReview { get; set; }
+        public double? MedieReview { get; set; }
         public List<string> Images { get; set; } = new List<string>();
         public double Distance { get; set; } = 0;
         public string? Website { get; set; }
